Разобьёт `main.js` на модули
